kafka-server/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/dev.in
#
astroid==2.12.10
    # via pylint
attrs==22.1.0
    # via
    #   -c requirements/prod.txt
    #   pytest
bandit==1.7.4
    # via -r requirements/dev.in
black==22.8.0
    # via -r requirements/dev.in
certifi==2022.9.24
    # via
    #   -c requirements/prod.txt
    #   requests
charset-normalizer==2.1.1
    # via
    #   -c requirements/prod.txt
    #   requests
click==8.0.4
    # via
    #   -c requirements/prod.txt
    #   black
    #   nltk
    #   safety
contourpy==1.0.5
    # via
    #   -c requirements/prod.txt
    #   matplotlib
cycler==0.11.0
    # via
    #   -c requirements/prod.txt
    #   matplotlib
dill==0.3.5.1
    # via pylint
dparse==0.6.2
    # via safety
drawnow==0.72.5
    # via -r requirements/dev.in
editdistance==0.6.0
    # via
    #   -c requirements/prod.txt
    #   -r requirements/dev.in
fonttools==4.37.4
    # via
    #   -c requirements/prod.txt
    #   matplotlib
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via bandit
idna==3.4
    # via
    #   -c requirements/prod.txt
    #   requests
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
itermplot==0.331
    # via -r requirements/dev.in
joblib==1.2.0
    # via
    #   -c requirements/prod.txt
    #   nltk
kiwisolver==1.4.4
    # via
    #   -c requirements/prod.txt
    #   matplotlib
lazy-object-proxy==1.7.1
    # via astroid
matplotlib==3.6.1
    # via
    #   -c requirements/prod.txt
    #   -r requirements/dev.in
    #   drawnow
    #   itermplot
mccabe==0.7.0
    # via pylint
mypy==0.981
    # via -r requirements/dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
nltk==3.7
    # via -r requirements/dev.in
numpy==1.23.3
    # via
    #   -c requirements/prod.txt
    #   contourpy
    #   itermplot
    #   matplotlib
    #   opencv-python
    #   pandas
    #   patsy
    #   scipy
    #   statsmodels
opencv-python==4.6.0.66
    # via -r requirements/dev.in
packaging==21.3
    # via
    #   -c requirements/prod.txt
    #   dparse
    #   matplotlib
    #   pytest
    #   safety
    #   statsmodels
pandas==1.5.0
    # via
    #   -c requirements/prod.txt
    #   statsmodels
pathspec==0.10.1
    # via black
patsy==0.5.3
    # via statsmodels
pbr==5.10.0
    # via stevedore
pillow==9.2.0
    # via
    #   -c requirements/prod.txt
    #   -r requirements/dev.in
    #   matplotlib
platformdirs==2.5.2
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycodestyle==2.9.1
    # via -r requirements/dev.in
pydocstyle==6.1.1
    # via -r requirements/dev.in
pylint==2.15.3
    # via -r requirements/dev.in
pyparsing==3.0.9
    # via
    #   -c requirements/prod.txt
    #   matplotlib
    #   packaging
pytest==7.1.3
    # via -r requirements/dev.in
python-dateutil==2.8.2
    # via
    #   -c requirements/prod.txt
    #   matplotlib
    #   pandas
python-dotenv==0.21.0
    # via -r requirements/dev.in
pytz==2022.2.1
    # via
    #   -c requirements/prod.txt
    #   pandas
pyyaml==6.0
    # via
    #   -c requirements/prod.txt
    #   -r requirements/dev.in
    #   bandit
regex==2022.9.13
    # via nltk
requests==2.28.1
    # via
    #   -c requirements/prod.txt
    #   safety
ruamel-yaml==0.17.21
    # via safety
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
safety==2.2.0
    # via -r requirements/dev.in
scipy==1.9.1
    # via
    #   -c requirements/prod.txt
    #   -r requirements/dev.in
    #   statsmodels
six==1.16.0
    # via
    #   -c requirements/prod.txt
    #   itermplot
    #   patsy
    #   python-dateutil
smmap==5.0.0
    # via gitdb
snowballstemmer==2.2.0
    # via pydocstyle
statsmodels==0.13.5
    # via -r requirements/dev.in
stevedore==4.0.0
    # via bandit
toml==0.10.2
    # via
    #   -c requirements/prod.txt
    #   dparse
tomli==2.0.1
    # via
    #   -c requirements/prod.txt
    #   black
    #   mypy
    #   pylint
    #   pytest
tomlkit==0.11.5
    # via pylint
tqdm==4.64.1
    # via
    #   -c requirements/prod.txt
    #   nltk
typing-extensions==4.3.0
    # via
    #   -c requirements/prod.txt
    #   astroid
    #   black
    #   mypy
    #   pylint
urllib3==1.26.12
    # via
    #   -c requirements/prod.txt
    #   requests
wrapt==1.14.1
    # via
    #   -c requirements/prod.txt
    #   astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
